backend/requirements.txt
flask
python-dotenv
google-generativeai
waitress
flask-cors
PyMuPDF